requirements.txt
certifi==2018.11.29
chardet==3.0.4
cmsplugin-filer==1.1.3
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.17
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.5.3
django-filer==1.4.1
django-filter==2.1.0
django-formtools==2.1
django-ipware==2.1.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-polymorphic==2.0.3
django-sekizai==0.10.0
django-treebeard==4.3
djangocms-admin-style==1.2.9
djangocms-attributes-field==1.0.0
djangocms-bootstrap4==1.3.1
djangocms-column==1.9.0
djangocms-file==2.2.0
djangocms-forms==0.2.5
djangocms-googlemap==1.2.0
djangocms-history==1.0.0
djangocms-icon==1.2.0
djangocms-installer==1.0.2
djangocms-link==2.3.1
djangocms-picture==2.1.3
djangocms-slider==0.2.5
djangocms-snippet==2.1.0
djangocms-style==2.1.0
djangocms-text-ckeditor==3.7.0
djangocms-video==2.1.1
djangorestframework==3.9.1
easy-thumbnails==2.5
et-xmlfile==1.0.1
hashids==1.2.0
html5lib==1.0.1
idna==2.8
jdcal==1.4
jsonfield==2.0.2
Markdown==3.0.1
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.12.0
sorl-thumbnail==12.5.0
tablib==0.12.1
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.24.1
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
